requirements.txt
alembic==1.0.9
boto3==1.9.190
botocore==1.12.190
click==6.7
docutils==0.14
Flask==1.0.2
Flask-Cors==3.0.8
Flask-Migrate==2.4.0
Flask-Script==2.0.6
Flask-SocketIO==4.1.0
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
gunicorn==19.9.0
itsdangerous==0.24
Jinja2==2.10.1
jmespath==0.9.4
Mako==1.0.9
MarkupSafe==1.0
psycopg2==2.8.2
PyJWT==1.7.1
python-dateutil==2.8.0
python-editor==1.0.4
python-engineio==3.8.2.post1
python-socketio==4.2.0
s3transfer==0.2.1
six==1.12.0
SQLAlchemy==1.3.3
urllib3==1.25.3
Werkzeug==0.14.1
WTForms==2.2.1
WTForms-JSON==0.3.3

